+++ src/main/java/ru/neoflex/cosmeticsStore/controllers/CustomersController.java
@@ -1,6 +1,6 @@
 package ru.neoflex.cosmeticsStore.controllers;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import ru.neoflex.cosmeticsStore.dto.CustomersDTO;
@@ -12,17 +12,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/customers")
 public class CustomersController {
 
     private final CustomersService customersService;
     private final MappingUtils mappingUtils;
-
-    @Autowired
-    public CustomersController(CustomersService customersService, MappingUtils mappingUtils) {
-        this.customersService = customersService;
-        this.mappingUtils = mappingUtils;
-    }
 
     @GetMapping
     public List<CustomersDTO> getAllCustomers() {
