+++ src/main/java/ru/neoflex/cosmeticsStore/dto/CustomersDTO.java
@@ -1,6 +1,12 @@
 package ru.neoflex.cosmeticsStore.dto;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
 public class CustomersDTO {
+
     Long id;
     String firstName;
     String lastName;
@@ -10,43 +16,4 @@ public class CustomersDTO {
     public CustomersDTO() {
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPhone() {
-        return phoneNumber;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setPhone(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
 }
