+++ src/main/java/ru/neoflex/cosmeticsStore/dto/OrdersDTO.java
@@ -1,7 +1,12 @@
 package ru.neoflex.cosmeticsStore.dto;
 
+import lombok.Getter;
+import lombok.Setter;
+
 import ru.neoflex.cosmeticsStore.entities.Customers;
 
+@Setter
+@Getter
 public class OrdersDTO {
 
     private Long id;
@@ -13,43 +18,4 @@ public class OrdersDTO {
     public OrdersDTO () {
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public Customers getCustomers() {
-        return customers;
-    }
-
-    public String getOrderDate() {
-        return orderDate;
-    }
-
-    public Integer getTotalAmount() {
-        return totalAmount;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setCustomers(Customers customers) {
-        this.customers = customers;
-    }
-
-    public void setOrderDate(String orderDate) {
-        this.orderDate = orderDate;
-    }
-
-    public void setTotalAmount(Integer totalAmount) {
-        this.totalAmount = totalAmount;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
 }
