+++ src/main/java/ru/neoflex/cosmeticsStore/entities/Gender.java
@@ -7,6 +7,11 @@ import jakarta.persistence.Column;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
 @Entity
 @Table(name = "gender")
 public class Gender {
@@ -20,22 +25,6 @@ public class Gender {
     private String description;
 
     public Gender () {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
     }
 
     @Override
