+++ src/main/java/ru/neoflex/cosmeticsStore/entities/Orders.java
@@ -9,6 +9,11 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
 @Entity
 @Table(name = "orders")
 public class Orders {
@@ -32,46 +37,6 @@ public class Orders {
     private String status;
 
     public Orders () {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public Customers getCustomers() {
-        return customers;
-    }
-
-    public String getOrderDate() {
-        return orderDate;
-    }
-
-    public Integer getTotalAmount() {
-        return totalAmount;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setCustomers(Customers customers) {
-        this.customers = customers;
-    }
-
-    public void setOrderDate(String orderDate) {
-        this.orderDate = orderDate;
-    }
-
-    public void setTotalAmount(Integer totalAmount) {
-        this.totalAmount = totalAmount;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
     }
 
     public String toString() {
