+++ src/main/java/ru/neoflex/cosmeticsStore/services/CustomersService.java
@@ -1,39 +1,17 @@
 package ru.neoflex.cosmeticsStore.services;
 
 import ru.neoflex.cosmeticsStore.entities.Customers;
-import ru.neoflex.cosmeticsStore.repository.CustomersRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class CustomersService {
+public interface CustomersService {
 
-    private final CustomersRepository customersRepository;
+    public List<Customers> getAllCustomers();
+    public Customers getCustomersById(Long id);
+    public Customers createCustomers(Customers customers);
+    public Customers updateCustomers(Customers customers);
+    public void deleteCustomers(Long id);
 
-    @Autowired
-    public CustomersService(CustomersRepository customersRepository) {
-        this.customersRepository = customersRepository;
-    }
-
-    public List<Customers> getAllCustomers() {
-        return customersRepository.findAll();
-    }
-
-    public Customers getCustomersById(Long id) {
-        return customersRepository.findById(id).orElse(null);
-    }
-
-    public Customers createCustomers(Customers customers) {
-        return customersRepository.save(customers);
-    }
-
-    public Customers updateCustomers(Customers customers) {
-        return customersRepository.save(customers);
-    }
-
-    public void deleteCustomers(Long id) {
-        customersRepository.deleteById(id);
-    }
 }
