+++ src/main/java/ru/neoflex/cosmeticsStore/services/ProductsService.java
@@ -1,39 +1,17 @@
 package ru.neoflex.cosmeticsStore.services;
 
 import ru.neoflex.cosmeticsStore.entities.Products;
-import ru.neoflex.cosmeticsStore.repository.ProductsRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class ProductsService {
+public interface ProductsService {
 
-    private final ProductsRepository productsRepository;
+    public List<Products> getAllProducts();
+    public Products getProductsById(Long id);
+    public Products createProducts(Products products);
+    public Products updateProducts(Products products);
+    public void deleteProducts(Long id);
 
-    @Autowired
-    public ProductsService(ProductsRepository ordersRepository) {
-        this.productsRepository = ordersRepository;
-    }
-
-    public List<Products> getAllProducts() {
-        return productsRepository.findAll();
-    }
-
-    public Products getProductsById(Long id) {
-        return productsRepository.findById(id).orElse(null);
-    }
-
-    public Products createProducts(Products products) {
-        return productsRepository.save(products);
-    }
-
-    public Products updateProducts(Products products) {
-        return productsRepository.save(products);
-    }
-
-    public void deleteProducts(Long id) {
-        productsRepository.deleteById(id);
-    }
 }
