+++ src/main/resources/application.yaml
@@ -1,0 +1,18 @@
+spring:
+  application:
+    name: CosmeticsStore
+  datasource:
+    driverClassName: org.postgresql.Driver
+    url: jdbc:postgresql://localhost:5432/CosmeticsStore
+    username: root
+    password: KOOLnasty25
+
+  jpa:
+    hibernate:
+      ddl-auto: create
+    generate-ddl: true
+    show-sql: false
+    database: postgresql
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
